Implementado os CRUDS e as mensagens de alerta

+++ loja-front/pages/cadastros/categoria/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import Sidebar from "@/components/Sidebar";
 import styles from "../../../styles/estado.module.css";
+import Swal from "sweetalert2";
 
 import axios from "axios";
 import { useState, useEffect } from "react";
@@ -36,20 +37,39 @@ const Categoria = () => {
   function handleSubmit(e) {
     e.preventDefault();
     if (categoria.id === undefined) {
-      axios.post("http://localhost:8080/api/categoria/", categoria).then((result) => {
-        setAtualizar(result);
-      });
+      axios.post("http://localhost:8080/api/categoria/", categoria)
+        .then(() => {return new Swal("Sucesso", "Categoria Cadastrada com Sucesso!", "success")}).then((result) => {
+          setAtualizar(result)
+        })
     } else {
-      axios.put("http://localhost:8080/api/categoria/", categoria).then((result) => {
-        setAtualizar(result);
+      axios.put("http://localhost:8080/api/categoria/", categoria)
+      .then(() => {return new Swal("Sucesso", "Dados Editados com Sucesso!", "success")}).then((result) => {
+        setAtualizar(result)
       });
     }
     limpar();
   }
 
   function excluir(id) {
-    axios.delete("http://localhost:8080/api/categoria/" + id).then(result => {
-        setAtualizar(result);
+    Swal.fire({
+      title: 'Você tem Certeza?',
+      text: "Essa ação não poderá ser desfeita!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Excluir'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios.delete("http://localhost:8080/api/categoria/" + id).then(result => {
+          Swal.fire(
+            'Excluído!',
+            'A Categoria foi excluída com sucesso!.',
+            'success'
+          )
+          setAtualizar(result)
+        })
+      }
     })
   }
